consolidate Stateful interface
