+++ src/model.ts
@@ -2,19 +2,12 @@ import type Public from '../types';
 
 import { createBindings } from './binding';
 import { useLookup } from './context';
-import { Controller } from './controller';
-import { useModel, useLazy, usePassive, useSubscriber, useWatcher } from './hooks';
-import { LOCAL, Subscriber } from './subscriber';
+import { CONTROL, Controller, Stateful } from './controller';
+import { useLazy, useModel, usePassive, useSubscriber, useWatcher } from './hooks';
 import { define, defineLazy, entries, fn, getPrototypeOf } from './util';
 
-export const CONTROL = Symbol("controller");
-export type Stateful = { [CONTROL]: Controller };
-
-export interface Model extends Public {};
+export interface Model extends Public, Stateful {};
 export class Model {
-  [CONTROL]: Controller;
-  [LOCAL]?: Subscriber;
-  
   constructor(){
     const control = new Controller(this);
 
@@ -74,9 +67,9 @@ export class Model {
     const instance: InstanceOf<T> = 
       new (this as any)(...args);
 
-    // evaluating CONTROL prevents compiler from dropping it.
-    // Triggering get required to properly initialize state.
-    return instance[CONTROL] && instance;
+    Controller.get(instance);
+
+    return instance;
   }
 
   static use(...args: any[]){
